CREDITS UPDATE
After a lot of hard work, fine tuning, and painful debugging, I bring
you this: THE FULLY WORKING CREDITS UPDATE!!!!!!!!! It is planned to
work with whatever comes and can easily be configured to come after the
statistics screen. On top of that, these credits give our game a much
nicer feel.